Resolve S10P31A106-359 "Be/members/"

See merge request s10-final/S10P31A106!93

+++ backend/src/main/java/life/bareun/diary/member/dto/MemberPracticedHabitDto.java
@@ -3,7 +3,7 @@ package life.bareun.diary.member.dto;
 
 public record MemberPracticedHabitDto(
     String habit,
-    int value
+    long value
 ) {
 
 }
